+++ CHANGELOG.md
@@ -1,1 +1,4 @@
+- Added more helpful error messages for the Firebase Hosting GitHub Action (#5749)
+- Upgrade Firestore emulator to 1.17.4
+- Web Frameworks will no longer try to deploy unsupported versions of NodeJS to Cloud Functions (#5733)
 - Adds function emulator support for v2 firestore triggers (#5685).